implement basic function call execution

+++ s-impl/src/main/java/mta/computional/slanguage/simpl/program/SProgramRunnerImpl.java
@@ -35,7 +35,7 @@ public class SProgramRunnerImpl implements SProgramRunner {
     }
 
     @Override
-    public long run(Integer... input) {
+    public long run(Long... input) {
 
         // empty program edge case
         if (program.length() == 0) {
