+++ s-model/src/main/java/mta/computional/slanguage/smodel/api/program/SProgramRunner.java
@@ -2,5 +2,5 @@ package mta.computional.slanguage.smodel.api.program;
 
 public interface SProgramRunner extends ExecutionContext {
 
-    long run(Integer... input);
+    long run(Long... input);
 }
